Clean Architecture Model
