--- .github/workflows/dotnetcore.yml
@@ -1,17 +1,0 @@
-name: .NET Core
-
-on: [push]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Setup .NET Core
-      uses: actions/setup-dotnet@v2
-      with:
-        dotnet-version: '7.0.x'
-    - name: Build with dotnet
-      run: dotnet build --configuration Release
