+++ SampleWebApiAspNetCore.Domain/DataTransferObjects/Dtos/FoodCreateDto.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SampleWebApiAspNetCore.Domain.DataTransferObjects.Dtos
+{
+    public class FoodCreateDto
+    {
+        [Required]
+        public string? Name { get; set; }
+        public string? Type { get; set; }
+        public int Calories { get; set; }
+        public DateTime Created { get; set; }
+    }
+}
