+++ SampleWebApiAspNetCore.Domain/DataTransferObjects/Dtos/FoodDto.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SampleWebApiAspNetCore.Domain.DataTransferObjects.Dtos
+{
+    public class FoodDto
+    {
+        public int Id { get; set; }
+        public string? Name { get; set; }
+        public string? Type { get; set; }
+        public int Calories { get; set; }
+        public DateTime Created { get; set; }
+    }
+}
