+++ SampleWebApiAspNetCore.Service/Helpers/DynamicExtensions.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Dynamic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SampleWebApiAspNetCore.Service.Helpers
+{
+    public static class DynamicExtensions
+    {
+        public static dynamic ToDynamic(this object value)
+        {
+            IDictionary<string, object> expando = new ExpandoObject();
+
+            foreach (PropertyDescriptor property in TypeDescriptor.GetProperties(value.GetType()))
+            {
+                expando.Add(property.Name, property.GetValue(value));
+            }
+
+            return expando as ExpandoObject;
+        }
+    }
+}
