+++ SampleWebApiAspNetCore.Service/Helpers/ExceptionExtension.cs
@@ -1,0 +1,30 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.Logging;
+using Microsoft.AspNetCore.Diagnostics;
+using Microsoft.AspNetCore.Http;
+
+namespace SampleWebApiAspNetCore.Service.Helpers
+{
+    public static class ExceptionExtension
+    {
+        public static void AddProductionExceptionHandling(this IApplicationBuilder app, ILoggerFactory loggerFactory)
+        {
+            app.UseExceptionHandler(errorApp =>
+            {
+                errorApp.Run(async context =>
+                {
+                    context.Response.StatusCode = 500;
+                    context.Response.ContentType = "text/plain";
+                    var errorFeature = context.Features.Get<IExceptionHandlerFeature>();
+                    if (errorFeature != null)
+                    {
+                        var logger = loggerFactory.CreateLogger("Global exception logger");
+                        logger.LogError(500, errorFeature.Error, errorFeature.Error.Message);
+                    }
+
+                    await context.Response.WriteAsync("There was an error");
+                });
+            });
+        }
+    }
+}
