+++ SampleWebApiAspNetCore.Service/Helpers/SeedDataExtension.cs
@@ -1,0 +1,26 @@
+﻿using Microsoft.AspNetCore.Builder;
+using Microsoft.Extensions.DependencyInjection;
+using SampleWebApiAspNetCore.Data;
+using SampleWebApiAspNetCore.Service.Services.Interfaces;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SampleWebApiAspNetCore.Service.Helpers
+{
+    public static class SeedDataExtension
+    {
+        public static void SeedData(this WebApplication app)
+        {
+            using (var scope = app.Services.CreateScope())
+            {
+                var dbContext = scope.ServiceProvider.GetRequiredService<FoodDbContext>();
+                var seedDataService = scope.ServiceProvider.GetRequiredService<ISeedDataService>();
+
+                seedDataService.Initialize(dbContext);
+            }
+        }
+    }
+}
