+++ SampleWebApiAspNetCore.Service/Helpers/VersioningExtension.cs
@@ -1,0 +1,35 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Versioning;
+using Microsoft.Extensions.DependencyInjection;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SampleWebApiAspNetCore.Service.Helpers
+{
+    public static class VersioningExtension
+    {
+        public static void AddVersioning(this IServiceCollection services)
+        {
+            services.AddApiVersioning(
+               config =>
+               {
+                   config.ReportApiVersions = true;
+                   config.AssumeDefaultVersionWhenUnspecified = true;
+                   config.DefaultApiVersion = new ApiVersion(1, 0);
+                   config.ApiVersionReader = ApiVersionReader.Combine(new UrlSegmentApiVersionReader(),
+                                                                                        new HeaderApiVersionReader("x-api-version"),
+                                                                                        new MediaTypeApiVersionReader("x-api-version"));
+               });
+           services.AddApiVersioning(x =>
+            {
+                x.DefaultApiVersion = new ApiVersion(1, 0);
+                x.AssumeDefaultVersionWhenUnspecified = true;
+                x.ReportApiVersions = true;
+                //x.ApiVersionReader = new HeaderApiVersionReader("x-api-version");  
+            });
+        }
+    }
+}
