+++ SampleWebApiAspNetCore.Service/Services/Interfaces/ISeedDataService.cs
@@ -1,0 +1,9 @@
+﻿using SampleWebApiAspNetCore.Data;
+
+namespace SampleWebApiAspNetCore.Service.Services.Interfaces
+{
+    public interface ISeedDataService
+    {
+        void Initialize(FoodDbContext context);
+    }
+}
