--- SampleWebApiAspNetCore/Dtos/FoodUpdateDto.cs
@@ -1,11 +1,0 @@
-﻿
-namespace SampleWebApiAspNetCore.Dtos
-{
-    public class FoodUpdateDto
-    {
-        public string? Name { get; set; }
-        public int Calories { get; set; }
-        public string? Type { get; set; }
-        public DateTime Created { get; set; }
-    }
-}
