--- SampleWebApiAspNetCore/Helpers/CorsExtension.cs
@@ -1,22 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace SampleWebApiAspNetCore.Helpers
-{
-    public static class CorsExtension
-    {
-        public static void AddCustomCors(this IServiceCollection services, string policyName)
-        {
-            services.AddCors(options =>
-            {
-                options.AddPolicy(policyName,
-                    builder =>
-                    {
-                        builder
-                            .AllowAnyOrigin()
-                            .AllowAnyHeader()
-                            .AllowAnyMethod();
-                    });
-            });
-        }
-    }
-}
