--- SampleWebApiAspNetCore/Helpers/ExceptionExtension.cs
@@ -1,27 +1,0 @@
-﻿using Microsoft.AspNetCore.Diagnostics;
-
-namespace SampleWebApiAspNetCore.Helpers
-{
-    public static class ExceptionExtension
-    {
-        public static void AddProductionExceptionHandling(this IApplicationBuilder app, ILoggerFactory loggerFactory)
-        {
-            app.UseExceptionHandler(errorApp =>
-            {
-                errorApp.Run(async context =>
-                {
-                    context.Response.StatusCode = 500;
-                    context.Response.ContentType = "text/plain";
-                    var errorFeature = context.Features.Get<IExceptionHandlerFeature>();
-                    if (errorFeature != null)
-                    {
-                        var logger = loggerFactory.CreateLogger("Global exception logger");
-                        logger.LogError(500, errorFeature.Error, errorFeature.Error.Message);
-                    }
-
-                    await context.Response.WriteAsync("There was an error");
-                });
-            });
-        }
-    }
-}
