--- SampleWebApiAspNetCore/Helpers/SeedDataExtension.cs
@@ -1,19 +1,0 @@
-﻿using SampleWebApiAspNetCore.Repositories;
-using SampleWebApiAspNetCore.Services;
-
-namespace SampleWebApiAspNetCore.Helpers
-{
-    public static class SeedDataExtension
-    {
-        public static void SeedData(this WebApplication app)
-        {
-            using (var scope = app.Services.CreateScope())
-            {
-                var dbContext = scope.ServiceProvider.GetRequiredService<FoodDbContext>();
-                var seedDataService = scope.ServiceProvider.GetRequiredService<ISeedDataService>();
-
-                seedDataService.Initialize(dbContext);
-            }
-        }
-    }
-}
