--- SampleWebApiAspNetCore/MappingProfiles/FoodMappings.cs
@@ -1,16 +1,0 @@
-﻿using AutoMapper;
-using SampleWebApiAspNetCore.Dtos;
-using SampleWebApiAspNetCore.Entities;
-
-namespace SampleWebApiAspNetCore.MappingProfiles
-{
-    public class FoodMappings : Profile
-    {
-        public FoodMappings()
-        {
-            CreateMap<FoodEntity, FoodDto>().ReverseMap();
-            CreateMap<FoodEntity, FoodUpdateDto>().ReverseMap();
-            CreateMap<FoodEntity, FoodCreateDto>().ReverseMap();
-        }
-    }
-}
