--- SampleWebApiAspNetCore/Repositories/FoodDbContext.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using SampleWebApiAspNetCore.Entities;
-
-namespace SampleWebApiAspNetCore.Repositories
-{
-    public class FoodDbContext : DbContext
-    {
-        public FoodDbContext(DbContextOptions<FoodDbContext> options)
-            : base(options)
-        {
-        }
-
-        public DbSet<FoodEntity> FoodItems { get; set; } = null!;
-    }
-}
