--- SampleWebApiAspNetCore/Repositories/IFoodRepository.cs
@@ -1,17 +1,0 @@
-﻿using SampleWebApiAspNetCore.Entities;
-using SampleWebApiAspNetCore.Models;
-
-namespace SampleWebApiAspNetCore.Repositories
-{
-    public interface IFoodRepository
-    {
-        FoodEntity GetSingle(int id);
-        void Add(FoodEntity item);
-        void Delete(int id);
-        FoodEntity Update(int id, FoodEntity item);
-        IQueryable<FoodEntity> GetAll(QueryParameters queryParameters);
-        ICollection<FoodEntity> GetRandomMeal();
-        int Count();
-        bool Save();
-    }
-}
