--- SampleWebApiAspNetCore/Services/ISeedDataService.cs
@@ -1,9 +1,0 @@
-﻿using SampleWebApiAspNetCore.Repositories;
-
-namespace SampleWebApiAspNetCore.Services
-{
-    public interface ISeedDataService
-    {
-        void Initialize(FoodDbContext context);
-    }
-}
