--- SampleWebApiAspNetCore/Services/SeedDataService.cs
@@ -1,18 +1,0 @@
-﻿using SampleWebApiAspNetCore.Entities;
-using SampleWebApiAspNetCore.Repositories;
-
-namespace SampleWebApiAspNetCore.Services
-{
-    public class SeedDataService : ISeedDataService
-    {
-        public void Initialize(FoodDbContext context)
-        {
-            context.FoodItems.Add(new FoodEntity() { Calories = 1000, Type = "Starter", Name = "Lasagne", Created = DateTime.Now });
-            context.FoodItems.Add(new FoodEntity() { Calories = 1100, Type = "Main", Name = "Hamburger", Created = DateTime.Now });
-            context.FoodItems.Add(new FoodEntity() { Calories = 1200, Type = "Dessert", Name = "Spaghetti", Created = DateTime.Now });
-            context.FoodItems.Add(new FoodEntity() { Calories = 1300, Type = "Starter", Name = "Pizza", Created = DateTime.Now });
-
-            context.SaveChanges();
-        }
-    }
-}
