+++ SamplwWebApiAspNetCore.Data/Repositories/Interfaces/IFoodRepository.cs
@@ -1,0 +1,22 @@
+﻿using SampleWebApiAspNetCore.Domain.Entities;
+using SampleWebApiAspNetCore.Domain.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SampleWebApiAspNetCore.Data.Repositories.Interfaces
+{
+    public interface IFoodRepository
+    {
+        FoodEntity GetSingle(int id);
+        void Add(FoodEntity item);
+        void Delete(int id);
+        FoodEntity Update(int id, FoodEntity item);
+        IQueryable<FoodEntity> GetAll(QueryParameters queryParameters);
+        ICollection<FoodEntity> GetRandomMeal();
+        int Count();
+        bool Save();
+    }
+}
